rename environment store filename remove type in axiom-store
