+++ npm/react/libs/core/src/store/environment-store.ts
