+++ npm/react/libs/core/src/tests/environment-store.spec.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import type { Environment } from '../models/environment';
-import { useEnvironmentStore } from '../store/environment.store';
+import { useEnvironmentStore } from '../store/environment-store';
 
 const makeEnvironment = (
   overrides: Partial<Environment> = {},
